Minimize chance of direction being less than 1.0
The child_dir that is computed below and fed here, recursively, can sometimes be less than 1.0, due to precision problems. We essentially set an epsilon of 0.0001 here to force vectors near the +Z or -Z axis to be treated as exactly these axes.

Another fix would be to normalize child_dir instead of dividing by "scale," but that is less efficient and I trust it a bit less to always guarantee the z component is 1 or -1. Epsilon seems safer in this case, where the spheres are normally rotated a large amount each time.